Add documentation

+++ game.c
@@ -4,8 +4,10 @@
 #include <stdlib.h>
 #include <string.h>
 
+/// Valid symbols to render a cell.
 const char symbols[3] = {'-', 'X', 'O'};
 
+/// Available patterns to match to check for a win.
 enum patterns {
     HORIZONTAL,
     VERTICAL,
@@ -14,15 +16,22 @@ enum patterns {
     MAX_PATTERN
 };
 
+/// Get a cell's value in the game's field.
 unsigned char game_get_field_value(ttt_game *game, unsigned long long x, unsigned long long y);
+/// Set a cell's value in the game's field.
 void          game_set_field_value(ttt_game *game, unsigned long long x, unsigned long long y, unsigned char value);
 
+/// Render the playing field.
 void render_field(ttt_game *game);
+/// Render a single row of the playing field.
 void render_field_line(unsigned char *field_line, unsigned long long line_size);
+/// Render a separator between two rows.
 void render_field_separator(unsigned long long line_size);
 
+/// Read coordinate input from stdin.
 unsigned long long get_coordinate_input(const char *name, unsigned long long limit);
 
+/// Check if somebody won the game.
 int check_win(ttt_game *game);
 
 int game_create(ttt_game *game, int argc, const char **argv) {
@@ -61,18 +70,20 @@ void game_run(ttt_game *game) {
     while (game->turns_played < game->cell_count) {
         printf("%c's turn:\n", symbols[cur_player]);
         render_field(game);
-        int r = get_coordinate_input("row", game->line_size);
-        int c = get_coordinate_input("column", game->line_size);
+        unsigned long long r = get_coordinate_input("row", game->line_size);
+        unsigned long long c = get_coordinate_input("column", game->line_size);
         if (game_get_field_value(game, c, r)) {
             printf("you cannot overwrite this tile\n");
             continue;
         }
         game_set_field_value(game, c, r, cur_player);
         if (check_win(game)) {
+            // if someone won, it must be the current player
             printf("player %c won the game\n", symbols[cur_player]);
             game->winner = cur_player;
             break;
         }
+        // alternate between 1 and 2
         cur_player ^= 3;
         ++game->turns_played;
     }
@@ -102,6 +113,7 @@ void render_field(ttt_game *game) {
 
 void render_field_line(unsigned char *field_line, unsigned long long line_size) {
     for (int j = 0; j < line_size; ++j) {
+        // field only stores index into the symbols array, not the actual symbols
         printf("|%c", symbols[field_line[j]]);
     }
     printf("|\n");
@@ -127,17 +139,20 @@ int check_win(ttt_game *game) {
     int matches[MAX_PATTERN + 1];
     matches[MAX_PATTERN] = 0;
     for (int i = 0; i < game->line_size; ++i) {
+        // check base element for each pattern
         matches[HORIZONTAL]  = game_get_field_value(game, 0, i);
         matches[VERTICAL]    = game_get_field_value(game, i, 0);
         matches[DIAGONAL_TD] = game_get_field_value(game, 0, 0);
         matches[DIAGONAL_BU] = game_get_field_value(game, 0, game->line_size - 1);
         for (int j = 1; j < game->line_size; ++j) {
+            // check next element in the pattern
             matches[HORIZONTAL] &= game_get_field_value(game, 0, i) == game_get_field_value(game, j, i);
             matches[VERTICAL] &= game_get_field_value(game, i, 0) == game_get_field_value(game, i, j);
             matches[DIAGONAL_TD] &= game_get_field_value(game, 0, 0) == game_get_field_value(game, j, j);
             matches[DIAGONAL_BU] &= game_get_field_value(game, 0, game->line_size - 1)
                                  == game_get_field_value(game, j, game->line_size - j - 1);
         }
+        // iff at least one pattern matched then matches[MAX_PATTERN] will not be 0
         for (int i = 0; i < MAX_PATTERN; ++i) {
             matches[MAX_PATTERN] |= matches[i];
         }
